+++ src/client/LangSelector.ts
@@ -239,8 +239,6 @@ export class LangSelector extends LitElement {
       "token-login",
     ];
 
-    document.title = this.translateText("main.title") ?? document.title;
-
     document.querySelectorAll("[data-i18n]").forEach((element) => {
       const key = element.getAttribute("data-i18n");
       if (key === null) return;